Use u64 for time

+++ kraken_sdk_rest/src/api/get_deposit_status.rs
@@ -47,7 +47,7 @@ pub struct DepositStatus {
     pub info: String,
     pub amount: String,
     pub fee: String,
-    pub time: u32,
+    pub time: u64,
     pub status: String,
     #[serde(rename(deserialize = "status-prop"))]
     pub status_prop: Option<String>,
